Merged PR 670314: Optionally use only local locations on unverified pins
The PR adds a flag to use only local locations for unverified pins because global locations can be overly optimistic because we never send trim events to the global store.

This PR is a repair item in order to mitigate the eviction storms in DM_PS01.

UPDATE:
An analysis even after reducing the lifetime of global entries in DM_PS01, in 60% of eviction storms the local information was better (less then 8 locations), but the global information was overly optimistic.

Related work items: #1985438