Merged PR 684515: Don't reuse server process when cache init fails
The L3 cache requires VPN or a PAT in order to auth. If you forget to connect to VPN, some amount of auth details get cached in the bxl.exe server process. It is annoying and undiscoverable that you have to cycle the bxl.exe process after you turn your VPN back on. This change makes it so the server process isn't reused if cache init fails.

Related work items: #2000530