Merged PR 685532: Fix RuntimeCacheMissAnalyzer when distributed source file hashing is enabled
When distributed source file hashing is enabled, the orchestrator might not have the hashes for all source files because the workers will hash the source files on demand. However, runtimecachemissanalyzer assumes that the orchestrator contains all source file hashes. It uses those hashes to compute the json for the weakfingerprint.

To mitigate this issue, I send the source file hashes as a part of ProcessStrongFingerprintComputationData.