Merged PR 684880: Rehash files in the SymbolDaemon if the existing hash type is not compatible
Rehash files in the SymbolDaemon if the existing hash type is not compatible.

In Office, we use murmur hash for local builds. In order to run automation for local builds, we need a build drop as well as a symbol drop so we are able to resolve call stacks during the automation runs.

There was work done to [rehash files for the drop](https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/commit/e5c60a0de09a64f9a33bc235b1ecb4972935a37e?refName=refs%2Fheads%2Fmaster) so the build drop would be available for automation, but it was never done for the symbol drop. This change adds the functionality to rehash files in the symbol drop so Office devs can create a symbol drop locally and automation can consume symbols from the symbol server to resolve call stacks during test execution

Related work items: #2002555