Merged PR 685974: Do not turn on AssumeCleanOutputs by default on CloudBuild
The problem manifests when cache convergence happens. This is when two pips run at the same time (on different builds) and one gets cached before the other one. When the one that finishes last is done, we discover that another pip beat it, so we converge. This means we align to the outputs that are already stored in the cache. So we clean up the outputs produced by the just executed pip and replay the outputs from what we find in the cache. But the clean-up part was messed up by this assumeCleanOutputs feature. So outputs were not being cleaned up and we were ending up with the just executed outputs on top of the cache replayed outputs. Since this pip is non-deterministic (the GUID is different every time), we had  two paths for the same content. The consumer pip was enumerating that directory, discovering both files and reading from both. But one of those files was a bogus one, which was not recognized as a true output, and therefore the DFA.