Merged PR 686095: [AnyBuild][BXL-Remoting] Ignore message processing failure when Detours is disabled
This will relieve rare and non-deterministic error in BXL remoting.

For BXL remoting, the pip is run under a sandbox-executor-wrapper. The wrapper itself is run under a sandbox (for enabling bindflt) but with Detours disabled.  Thus, it's safe to ignore reported file message from sandbox because they are not going to be used anyway.