Merged PR 687258: Avoid deploying MSBuild runtime dlls that are provided by the installed instance
We compile the MSBuild graph builder tool against some dlls in order to call the static graph APIs. But at runtime, we use the user-provided MSBuild instance, so we don't want to deploy any msbuild-related dll as runtime content.
It seems we added some other target frameworks for existing dlls as part of bxl runtime dependencies, and those were leaking into the msbuild graph builder tool. This PR makes sure that those dlls are not deployed