Merged PR 685994: Filterout inactive machines for global locations
This PR adds 2 flags:

* `TraceInactiveMachinesForGlobalLocations` to just trace inactive machines obtained from the global store.
* `FilterInactiveMachinesForGlobalLocations` to filter out inactive machines from the list of global locations.

When we added filtering logic for inactive machines we were very explicit not to filter out them for global locations because its possible that the machine changes the state from 'inactive' to 'active' adds the content to the global store and then in a second that content is requested by another machine. At that point the machine state still could be stale and we might fail the build.

The first flag will allow us to observe if this is a possible case or not.

The overall reason for this PR is that we noticed that in one of the stamps (MW_PS05) we got stuck with one location that GCS returns all the time for an inactive machine causing build failures for `bulidTest_CloudSign_MWPS05` queue.

Related work items: #2003953