Merged PR 687591: Produce an ubuntu-based sandbox
This change pushes to main the changes required to produce an Ubuntu-based sandbox to the current off band pipeline that produces the Linux sandbox package. This won't be needed after we start producing the sandbox in-build, but until that happens, this makes sure any changes in the sandbox that need publishing will be included in an Ubuntu-produced package.