Merged PR 688225: Use ActionBlockSlim in content hashing
**THIS LOOKS UGLY AND HACKY** but it works!

This PR removes the usage of 'ActionBlock' from the hashing library, but unfortunately, we can't directly just switch to 'ActionBlockSlim' by adding the reference to BuildXL.Utilities. The hashing library is a very widely used dll and adding any extra dependency would be consider a major breaking change.

So instead of adding dll -> dll dependency this PR adds the source reference dependency by using 'ActionBlockSlim.cs' in the hashing project by referencing that file manually. The other downside of this approach is that we have to change the type visibility to avoid conflicts.