Merged PR 687960: Upgrade RuntimeContracts to 0.4.0
This PR updates RuntimeContracts to 0.4.0 with the following important changes:

* All the assertions propagate caller expressions as string to make assertions more descriptive. For instance, `Contract.Requires(x != null)` will automatically embed `x != null` into a message.
* All the assertions support new interpolated string from C# 10 and allow lazy message construction. It means that now we don't need to use 'FluentAPI' and just create custom message with interpolated string syntax.
The following code is very efficient: `Contract.Assert(x > 0, $"x = {x}")`. The efficiency comes from the fact that the message is constructed only if the assertion is false and the exception will be thrown.

This PR also fixes cases when the messages were created eagerly causing excessive memory allocations that are now gone.

Related work items: #2007282