Merged PR 684968: Allow workers to actively join the distributed build
Add a `Hello` RPC that will be initiated by workers that know the location of the orchestrator when they launch the build. After learning of the worker's location, the regular attachment procedure is initiated. If we already know the workers from the command line, the RPC has no effect.

Also update the AdoBuildRunner to use this distribution model.