Merged PR 689696: Update BuildXL package to include copyfilerange patch for Linux sandbox
On newer kernel, copyfilerange no longer works cross-device. This is a known regression in the kernel.

The current BuildXL package has the Linux sandbox that doesn't contain this copyfilerange patch. AnyBuild requires this patch as it now has an updated Linux image for its agents. As a workaround, AnyBuild currently relies on two packages from BuildXL, i.e., buildxl package itself and runtime.ubuntu-linux-x64.buildxl. The latter is produced in an adhoc way.

With this PR, we can now rely solely on buildxl package.