Merged PR 689879: Stop using 'Contract.Check' in favor of regular Contract API
After upgrading 'RuntimeContracts' to version 0.4.0 we can retire usage of the fluent contracts API like `Contract.Check(x > 0)?.Requires("message")` in favor of regular API. The fluent API was added for pre-C#10 version to avoid message allocations for regular cases (when the contract is not fails).

This PR was created by applying the code fix.