Merged PR 688716: Adding a timeout for flushing messages on shutdown
This PR adds a timeout for flushing pending LLS events on shutdown.

This should mitigate an error pattern happening in BN_PS02 when due to reconciliation the events are stored locally and lost. They will be sent during reconciliation but it might take longer than the content lifetime in the global store causing the builds to fail.

Here is the test results for this feature in DM_S2:
![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/688716/attachments/image.png)

You can see that the operation context was canceled in the middle of sending the data, but the events were still successfully delivered.

Related work items: #2008812