Merged PR 690328: [Linux sandbox] Reset cached path for file descriptor on open-like interposings
We have a cache from file descriptor to path that we clear only when we observe a call to `close`. However, we can potentially miss one of these observations and, if, the file descriptor is reused, we might report an incorrect path. This was observed to happen in Midgard builds. This PR also clears the file descriptor cache every time we observe an operation that returns a new file descriptor (such as `open`), assuming that the underlying path will have changed.

If it's the case that we observe all the `close`s, then this wouldn't make a difference, but being conservative here is a lot better than missing observations.