Merged PR 691594: Use sorted locations when merging content locations for perf reasons
We've found that merging the content locations is quite expensive on the workers:
![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/691594/attachments/image.png)

The new version adds an ability to store location changes in sorted manner to merge them on the fly without deserializing them into the custom types.

The performance improvements is massive: 3-5-x performance improvements and 0 allocations (!).

![image (2).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/691594/attachments/image%20%282%29.png)

Related work items: #2012550