Merged PR 693567: Avoid caching file descriptors pointing to non-file artifacts
Avoid storing in the file descriptor table elements that point to non-file artifacts (sockets, pipes, etc). We currently don't detour socket/pipe creation, so we cannot invalidate socket/pip creation in the same way we do for files (check the linked bug for a description of the scenario with more details).

The change also removes some invalidations that were not necessary.

Related work items: #2015184