Merged PR 694444: Close newly loaded store if shutdown was requested
This PR fixes the issue that was causing a bunch of tests to crash when the rocksdb logging was enabled for metadata store.

Due to background processes, its possible to start `Load` method before the shutdown started, but finish the load when the shutdown is done.

In this case, the newly loaded store will stay opened forever (because the store doesn't have a finalizer) and likely cause a crash if the logging is on because the logger adapter used to have a finalizer that was cleaning up the delegates allocated for handling the logs on the managed side.