Merged PR 694918: Fix command line construction of Javascript pips with extra arguments
On unix, we need to join the script command and the extra arguments, because we want all of it to be part of the same escaped unit in the fragment. This makes the difference (having "npm execute build" as script command and "--extra-argument" as extra argument) between this:

```bash -c "npm execute build --extra-argument"```

 and this:

```bash -c "npm execute build" --extra-argument```

The second one would be incorrect, losing the extra argument. This problem was observed to happen in practice in OOUI builds, where the lage commands are `npm run build`:

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/694918/attachments/image.png)