Merged PR 695063: Address the case of a dynamic write on a statically declared source file
The case of a dynamic write into a statically declared type was incorrectly reported as a double write. This PR addresses that case by adding a specific violation for this case. Note: observe that a regular (static) write that clashes with a statically declared source file is already addressed by construction at pip graph building time.