Merged PR 695817: Avoid generating a core dump on pip cancellation for Linux
On Windows we only generate core dumps on timeouts, not on pip cancellation. The Linux implementation is slightly different, and we were doing that on pip cancellation and in particular when cancelling a pip due to resource exhaustion. In some scenarios producing the dump was taking several minutes, and we were not fast enough to cancel the pip before we run out of memory.

This PR makes sure we only produce dumps on timeout, so this should match the Windows behavior.

Tried to add some tests, but unfortunately the related tests (e.g. RetryPipOnHighMemoryUsage) don't work on Linux. I can create a bug/update the existing work items so we add something for this whenever we fix these tests. I don't think adding tests is super critical anyway.

Related work items: #2018806