Merged PR 696346: Upon CB timeout, trigger the cancellation token before logging internal error to avoid a small race
Because the error we log on CB timeout is a infrastructure failure, this will trigger the stop-on-first-internal-error logic, which is a bit misleading in this particular case. If the cancellation token is already triggered the schedule is terminating and that logic is skipped.

I'm not aware of issues with this race, but I think it's clearer to not log the message in this case.