Merged PR 696582: Trace checkpoint information in case of restore failure as well
We had an ICM that was caused by the GCS restore errors, but when the restore fails, we don't trace key information that might be already available at that point. This PR fixes this.

Here is an example of a trace without critical information:

```
ResilientGlobalCacheService.RestoreCheckpointAsync stop 16.0606ms. result=[Error=[(JsonException: '0x00' is an invalid start of a value. Path: $ | LineNumber: 0 | BytePositionInLine: 0.) (JsonReaderException: '0x00' is an invalid start of a value. LineNumber: 0 | BytePositionInLine: 0.)], Diagnostics=[StackTrace:
    at async Task<CheckpointState> BuildXL.Cache.ContentStore.Distributed.MetadataService.ResilientGlobalCacheService.RestoreCheckpointDatabaseAsync(OperationContext context) in
]]. CheckpointId=[] DbLogId=[0] StartReadLogId=[0] StartWriteLogId=[0]
```