Merged PR 697067: Make allow-list-to-dynamic-output logic option part of unsafe option
The enablement of allow-list logic to dynamic outputs affects cached results. When enabled, one can have less outputs (due to being filtered out by the allow-list). Thus, the enablement of the feature needs to be included in the fingerprint.

In this PR that logic is part of unsafe configuration. This PR also bumps up the fingerprint version.

By default, the logic is disabled, unless there's JavaScript or MsBuild resolver specified in the configuration.