Merged PR 697065: Ensure MsBuild resolver schedule (project, qualifier) once
The bug associated with this PR shows contract violations caused by the resolver trying to add the same project node more than once.

The violation shows that we try to add the same value pip to the graph.

There are two bugs fixed by this PR. First, initially, the value pip's key is constructed from project's filename. In a MsBuild repo, there could be multiple projects that have the same filename, e.g., dirs.proj. The fix for this bug is to use the relative path to the root instead of just the project's filename.

The second bug is the evaluation of the module representing the MsBuild resolver can be evaluated more than once. It happens when the DScript resolver tries to resolve a value first, and then followed by the MsBuild resolver itself trying to evaluate its module. Both eventually calls `EvaluateAllFilesAsync`. The fix for this bug is to dedupe the calls (only one call at a time), and only happen once; the loser simply gets the value computed by the winner.

This PR also fixes the caching issue of process construction helper. Previously, there is no cache hit because the used key contains a reference to GlobalProperties object that is never shared.

Related work items: #2023444