Merged PR 697563: Have an explicit DScript option to set DoNotApplyAllowListToDynamicOutputs
Do not automatically turn off DoNotApplyAllowListToDynamicOutputs for JS/MSBuild resolvers. Rationale: the presence of such resolver in the resolver list will sneakily affect the behavior of other potential DScript resolvers in the same list. Instead, make an explicit DScript option to set DoNotApplyAllowListToDynamicOutputs and for the time being, repos that need outputs to be considered in the allow list will have to explicitly set it