Merged PR 697820: [Linux] Fix reporting file descriptor access
To get a path from a file descriptor, we call `readlink` on `/proc/self/fd`. The call may fail, e.g., the fd is no longer valid. In this case we get an empty path. However, for getting policy and reporting, the sandbox asserts that the path must be an absolute path (starting with `/`); thus a crash.

Related work items: #2025230