Merged PR 698197: Don't fail if conhost.exe survives symbolD execution
One of the modes of SymbolD is running as a process pip. In this mode, symbolD calls into a library to extract ClientKeys from pdbs. Recently, we observed a scenario where upon symbolD exiting there is still a running child conhost.exe process. We don't spawn this process, and to the best of our understanding, the lib we are using is not spawning it either. Because there is a surviving child process, BXL logs an error.

It's likely some weirdness of .net. Since SymbolD exits with 0 exit code, we can clean all child processes with no risk. So, this PR configures symbolD process pips to ignore a child conhost.exe if it survived.

Related work items: #2025770