Merged PR 698451: Use timestamps to avoid races when computing dir fingerprints in minimal grap...
For computing directory fingerprints in OIP when minimal graph + alien files mode is on we need to enumerate the real file system in search of source files. This search wants to discard all outputs not present in the pip file system, but there is a race when an output from a non-related pip is produced and didn't have time to reach the output file system. In those cases, the fp computation may pick those up.

This PR tries to make sure we minimize those races. Creation timestamps are now used to understand whether a file was created after the engine started. Those files are interpreted as output files. Observe we go through this check after checking whether the file is a registered output, as a 'catch-all' check. A couple of related optimizations are also added.