Merged PR 699751: Propagate operation errno as part of the access report on Linux
This PR adds the errno into the access report whenever an interposed operation fails. This puts this functionality at par with the Windows sandbox.

The act of creating an access report and actually reporting it is split in two: an access report is created and the access check is performed. The interposed function is maybe checked for access and executed. The created report is sent afterwards including the errno in case of a failure.

The ptrace sandbox still does not do this. @<Pasindu Gunasekara 🍣> , I added TODOs in all the places where this should happen, but we need to figure out how to the get the return value and errno first.

Related work items: #2026687