Merged PR 699637: Don't extract nuspec file during nuget package download
NugetDownloader pip is currently extracting all (non-nuget) files during package download. We don't really need the .nuspec file (we download it and inspect it as a part of front-end analysis, but there is no clear need for this file afterwards). Nuget lib that we are using has custom logic that handles the extraction of nuspec files -- it extracts the nuspec file, reads the package Id from it, and uses that id as a file name. Because of this, casing of nuspec in the archive might not match the casing of the extracted file. On Windows, it's not a problem; however, on Linux where paths are case-sensitive, this leads to DFAs. Since we don't really need nuspec files, this PR removes them from extraction.

Related work items: #1999989