Merged PR 700973: Allow BuildXL to run without change journal
In case we need to run BuildXL inside Windows container.

From Windows container dev:
"No, we do not support the USN for the use cases of BuildXL.

The problem is that we merge the scratch filesystem and the layer filesystem to produce a merged view. We do not have the ability to keep the USN journals for both the scratch filesystem and layer filesystem synced so that if there are changes to either the scratch or the layer, we are able to show the correct virtualized USN view."