Merged PR 701010: Report the right pid on the linux sandbox
All reports had getpid() as the process id. But the issue is that fork() is reported on the parent process (since the parent process tracks childs), so getpid() is not the right id to report in that case. This was causing file accesses to get the wrong ReportedProcess on managed side (and as a side effect was triggering a lot of warnings about the process being null when checking allowlist entries against the image name).

Moreover, all exec* functions were also reporting the wrong pid since the report was happening before calling exec*. This reporting was actually not needed, since every time a new instance of the observer is spawned, we report a process creation anyway (with the right id).

@<Pasindu Gunasekara 🍣>, I think the change I did here on the exec* should also apply for the new ones you are trying to add in this in-flight PR you have open.