Merged PR 701421: Update minimal AnyBuild client requirement to 6f4f1d5d_20230206.7_275703
Office is still using BuildXL remoting to speed up their big builds. We've recently changed the remoting API, and the change is not backward compatible.

For developer builds, BuildXL tries to avoid re-installing AnyBuild when unnecessary, based on the minimal version requirement. So, developers may have stayed with the old local installation.