Merged PR 701482: TwoLevelCache can skip remote put if local already has content
As described in the in-code comment, this is meant as an optimization. As for the paging logic, I've seen cases where the `Task.WhenAll` call throws null exception because one of the tasks is null, which leads me to believe that somehow the page number calculation is wrong. To avoid doing the page number calculation altogether, I'm creating a list rather than an array. Less efficient, so if this is a critical path feel free to fight me about it.

Related work items: #2030893