Merged PR 701426: Linux sandbox sends all debug messages to bxl main log file
This matches the Windows behavior where all detours debug messages are sent to the main bxl log so the experience is consistent.
Additionally in many situations having a separate log is not easy to configure (e.g. unit tests and its different infrastructures).

Finally, we saw some weird hangs when reporting messages in the old way, so this might fix it.