Merged PR 701891: Use std::string instead of char* in the FilesCheckedForAccess set
`PCPathChar` amounts to `const char *`, and the values of these pointers are not right to store as the elements of the set: we want the actual strings to be there.

This is causing incorrect behavior in the TryRegisterPath method, which in turn makes us skip or duplicate some reports.

Related work items: #2031512