Merged PR 702015: [Linux sandbox] Fix invalid pid when running under AnyBuild container
When running under AnyBuild container with limited max pids, the pid in the report structure can be -1 because the process may not be created. The managed sandbox connection failed to parse the pid, and assumed that the pid is 0, so it tracks a process with pid 0. But then that process is never removed from the track set, causing the build to be non-terminating.

```
2023-02-13 17:16:57,683 [18] DEBUG (2d03cd9a) DetoursEventListener: CmdId: 2d03cd9a: Detours: Exited: False, StdOut: False, StdErr: False, Reports: False :: [ERROR]: Could not parse int from '-1' (errno: 0)
2023-02-13 17:16:57,683 [18] DEBUG (2d03cd9a) DetoursEventListener: CmdId: 2d03cd9a: Detours: Exited: False, StdOut: False, StdErr: False, Reports: False :: AddPid(0) :: added: True; size: 11
```

We restore the old behavior in the sense that if the reported pid is < 0, then we use the pid of the current process, which is obtained by calling `getpid()`.

This PR fixes AnyBuild broken builds, and allows AnyBuild to ingest new BuildXL.