Merged PR 701922: YAML pipeline for public validation in Rolling build
The intent is to have our rolling pipeline invoke this yaml pipeline (and embed it when we eventually move public over to yaml). This new public pipeline does not have access to any of the secrets that allow access to internal feeds. This will ensure there is no crosstalk between our internal and publicly available packages.