Merged PR 702258: Propagate mkdir return value in ptrace sandbox
Propagate mkdir return value in ptrace sandxbox. In order to test this, a new debug flag is added that enables to force the ptrace sandbox to always run. We could use this later for a better test infra for ptrace where we can reuse existing integration tests.

Related work items: #2028824