Merged PR 702243: Define CanonicalizedPathType as std::string rather than PCPathChar to avoid storing dangling pointers
Storing the result of std::string::c_str() is unsafe, as the pointer will point to an invalid location after the string that produced it is out of scope, destroyed of modified. We were storing some of these pointers in the form of CanonicalizedPathType in the PolicyResult, and then used those (possibly invalidated) pointers to report the operation paths, (possibly) leading to incorrect or duplicate reports. This PR makes CanonicalizedPathType a string instead of a PCPathChar (i.e. char *) to avoid this pitfall in the future.