Merged PR 702620: Fix mistargeted directory symlinks in opaque directories when preserved outputs is enabled
When PreserveOutputs is enabled for an entire opaque directory, BuildXL leaves the directory as-is and only checks the state of individual files during cache replay. When a file needs to be materialized, it creates all parent directories. However, if the state of the opaque directory is altered so that those parent directories exist as symlinks, BuildXL currently crashes.

The per-file materialization code does not have the context to know whether it is appropriate to create the target of the dangling directory symlink or to delete the symlink and create a real directory. So this is instead handled when BuildXL would have deleted the opaque directory. Now all opaque directories go through the cleanup process, but the cleanup logic targets ensuring relevant directory symlinks are in the correct state if preserveOutputs is enabled for the opaque directory.

Related work items: #1946137