Fix the Debugging Documentation
The debugging documentation currently doesn't list the namespace of the
functions it describes, causing developers to spend time reverse
engineering the debug object.
This commit adds the namespace to the documentation to make it clear to
developers that these functions need to be prefixed by the Debug
namespace when invoking them.