Merged PR 703804: Promote credscan detections from warnings to errors and refactoring changes related to CredentialScanner class
A new change has been added to PipGraphFragmentGenerator to handle an Assertion failure.
With the new implementation of an Interface for CredentialScanner and NoOp implmentation of that interface, credscan instance is never null when it is not enabled.
The Assert condition has been changed accordingly.

Tested this change by scheduling two Office builds.
https://cloudbuild.microsoft.com/build/d84db64b-cd98-439a-b362-e9da08be9b98
https://cloudbuild.microsoft.com/build/38a617cc-2dd8-41d5-936e-091a2d77e398

The PR has been approved already but reverted
Here is the previous PR.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/702242