Merged PR 703946: Don't FailFast when creating a WER report fails
When BuildXL crashes it creates a custom Windows Error Reporting report with some additional information beyond the standard CLR report. When there's a failure creating that custom WER report, the code currently takes down the application with an Environment.FailFast.

Since we primarily utilize telemetry for tracking crashes, this FailFast is preventing us from seeing a lot of crash data.