Merged PR 705053: [Linux sandbox] Do not rely on the process environment, which is volatile, to communicate the FAM path
The environment might be cleared at any point by the running process, making us lose the values if we're relying on an environment variable being set for it. Instead, store the path in the observer for subsequent uses.

The environment being cleared is not a theoretical problem, it was observed in practice, and it made us crash, see: #2036917

Related work items: #2036917