Merged PR 703227: Allow machine total ram to dynamically increase for ram forecasting
BuildXL currently snaps total machine memory (working set) at the beginning of the build. This is normally a safe assumption, but builds on machines that are VMs with dynamic memory on the host will vary the amount of assigned memory while the build happens. This change allows what is used for modeling resources to grow while the build is running.