Merged PR 705907: Report process creation for clone/fork on the child process
Reporting process creation for fork/clone on the parent has the problem that report lines from the child may non-deterministically arrive before the process creation report line. This causes that on managed side we don't have a process entry to associate those reports to, and therefore an allowlist that needs the current executable for a given access cannot be applied. This PR moves the reporting to the child.

Related work items: #2038280