Merged PR 706179: Merge ConcurrencyLimit and UnavailableSlots in build summary reporting
The Limiting factors summary at the end of the build had a heuristic for ConcurrencyLimit. That was calculated based on there being free resources on the machine and more work available to be run. Evolutions to the scheduler eventually codified this heuristic into the UnavailableSlots counter. This superseded the ConcurrencyLimit in precedence in ExecutionSampler. So ConcurrentyLimit was basically always getting logged at 0%.

I consolidated the two keeping the original ConcurrencyLimit terminology since that was used in the performance summary at the end of the build log.

Related work items: #2033907