Merged PR 706676: Handle pips that are too large to serialize to fingerprint store
The root cause of this crash was a pip with 8 million dependencies. That was unintended in spec generation and has been addressed, but I figured it was still good to have some form of handling in case that pops up again.

Originally I wanted to handle this in a more elegant way where we would keep fingerprint text that was truncated after a certain point. But that would not be valid json. Truncating cleanly while still allowing the JSON to be valid is more effort than it's worth so just going with logging a warning and moving on.

Related work items: #2036600