Merged PR 707064: Adding error handling to CreateFileStream() GraphAgnosticIncrementalSchedulingState.cs in to avoid crashing on failure during incremental scheduling.
Adding error handling to CreateFileStream method to avoid crashing on failure during incremental scheduling.
We do not want the build to crash or exit after the exception has occurred as the build proceeds further without any issues even if there is any error, hence a warning is logged here instead of an error.
If we track this call one of the inner methods CreateFileStream in FileSystem.Win.cs throws a BuildXLException
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal?path=/Public/Src/Utilities/Native/IO/Windows/FileSystem.Win.cs&version=GBmain&line=1985&lineEnd=1986&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents
hence I added a catch block to handle that exception but if we want to make it exclusive to IOException I can modify this.

Related work items: #2012530