Merged PR 706729: Fix a crash caused by a race during build cancellation
When we are approaching a CB timeout, we trigger a cancellation token. As a result, we cancel all running pips as well as cancel the post processing of all finished pips (even if they were successful). This pr fixes the assert, so we don't crash during build termination.

Related work items: #2028490